code polishing
git-svn-id: https://svn.code.sf.net/p/openuss/code/branches/solarplexus@1564 60046477-0c10-0410-bf4b-e552a4ff1768

+++ openuss-plexus/plexus/plexus-web/src/main/java/org/openuss/web/search/ExtendedSearchUtil.java
@@ -21,9 +21,9 @@ public class ExtendedSearchUtil {
 		return rb;
 	}
 	
-	public static String getResourceExtensionString(Integer searchScopeId){
+	public static String getResourceExtensionString(Long searchScopeId){
 		String extension = "";
-		switch(searchScopeId){
+		switch(searchScopeId.intValue()){
 			case Constants.EXTENDED_SEARCH_SCOPE_UNIVERSITIES:
 				extension = "univ";
 				break;
